fix: 修复验证时的 fps 警告和 BFloat16 可视化错误
1. fps 警告修复:
   - src/models/qwen3_vl/integration.py
   - src/models/qwen3_vl/sequence_packing.py
   添加 nframes 参数到 video content

2. BFloat16 可视化错误修复:
   - scripts/train.py: visualize_heatmap_predictions
   添加 .float() 转换避免 numpy 不支持 BFloat16

+++ scripts/train.py
@@ -342,7 +342,7 @@ def visualize_heatmap_predictions(
             axes[i, 1].set_title(f"GT Heatmap (max={gt_hm.max():.2f})")
             axes[i, 1].axis('off')
             
-            pred_hm = pred_heatmaps[i].detach().cpu().numpy()
+            pred_hm = pred_heatmaps[i].detach().float().cpu().numpy()  # float() 避免 BFloat16 错误
             pred_hm = np.clip(pred_hm, 0, 1)
             axes[i, 2].imshow(pred_hm, cmap='inferno', vmin=0, vmax=1)
             axes[i, 2].set_title(f"Pred Heatmap (max={pred_hm.max():.2f})")
